--- a/packages/graphback-codegen-schema/package.json
+++ b/packages/graphback-codegen-schema/package.json
@@ -25,6 +25,7 @@
   "devDependencies": {
     "@types/jest": "26.0.9",
     "@types/node": "12.12.54",
+    "dataloader": "2.0.0",
     "graphql": "15.3.0",
     "jest": "26.2.2",
     "rimraf": "3.0.2",