This removes the number of queries issued to the database

Fixes https://github.com/aerogear/graphback/issues/1304

## Reproducer.
Go to current [master](https://github.com/aerogear/graphback/commit/9e71c786e429984046a7f6dba022b0effd760ef6) and use the PostgresSQL template with the following model:

```graphql
""" @model """
type Note {
  id: ID!
  title: String!
  """
  @oneToMany(field: 'note')
  """
  comments: [Comment!]!
}

""" @model """
type Comment {
  id: ID!
  text: String
  note: Note!
}
```

## start the server

```shell
DEBUG=knex:query yarn develop
```
create a note which has three comments linked to it. 

If I issue a graphql query to retrieves *Notes* like below
```graphql
query {
  findNotes {
    items {
      id
      comments {
        id
      }
    }
  }
}

```

I am seeing only two db queries below thanks to dataloader, this is fine.

```log
querying object Note with filter undefined
  knex:query select * from "note" undefined +21s
  knex:query select * from "comment" where "noteId" in (?, ?) undefined +5ms
```

However, when I query for *Comments* with a query like 
```graphql
query {
  findComments {
    items {
      id,
        note {
          id
        }
    }
  }
}

```

I am seeing the following logs, notice the number of sql query to retrieve note. 
```log
knex:query select * from "comment" undefined +1m
  knex:query select * from "note" where "id" = ? limit ? undefined +13ms
  knex:query select * from "note" where "id" = ? limit ? undefined +5ms
  knex:query select * from "note" where "id" = ? limit ? undefined +3ms
  knex:query select * from "note" where "id" = ? limit ? undefined +0ms
....
```
I think this should be only one query, and not several ones. 

Apply this patch and you should see, and verify again to see that the number of queries has been reduced.